+++ src/roman_datamodels/stnode.py
@@ -22,6 +22,8 @@ from .validate import _check_type, ValidationWarning, _error_message
 import rad.resources
 from .stuserdict import STUserDict as UserDict
 
+from roman_datamodels.units import Unit
+
 if sys.version_info < (3, 9):
     import importlib_resources
 else:
@@ -34,6 +36,7 @@ __all__ = [
     "NODE_CLASSES",
     "CalLogs",
     "FileDate",
+    "Unit",
 ]
 
 
@@ -43,6 +46,9 @@ strict_validation = True
 # Needed due to enum bug with ASDF
 _VALID_ORIGIN = ["STSCI", "IPAC/SSC"]
 _VALID_TELESCOPE = ["ROMAN"]
+
+
+_UNIT_NODE_CLASSES_BY_TAG = {Unit._tag: Unit}
 
 
 def set_validate(value):
@@ -519,6 +525,30 @@ class TaggedScalarNodeConverter(Converter):
         return _SCALAR_NODE_CLASSES_BY_TAG[tag](node)
 
 
+class UnitConverter(Converter):
+    @property
+    def tags(self):
+        return list(_UNIT_NODE_CLASSES_BY_TAG.keys())
+
+    @property
+    def types(self):
+        return list(_UNIT_NODE_CLASSES_BY_TAG.values())
+
+    def to_yaml_tree(self, obj, tag, ctx):
+        import roman_datamodels.units as units
+
+        unit = obj.to_string()
+        if unit in units.ROMAN_UNIT_SYMBOLS:
+            return unit
+        else:
+            raise ValueError(f"Unit {unit} is not a valid Roman unit")
+
+    def from_yaml_tree(self, node, tag, ctx):
+        import roman_datamodels.units as units
+
+        return getattr(units, node)
+
+
 _DATAMODELS_MANIFEST_PATH = importlib_resources.files(
     rad.resources) / "manifests" / "datamodels-1.0.yaml"
 _DATAMODELS_MANIFEST = yaml.safe_load(_DATAMODELS_MANIFEST_PATH.read_bytes())
@@ -566,6 +596,8 @@ for tag in _DATAMODELS_MANIFEST["tags"]:
         _LIST_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
     elif tag["tag_uri"] in _SCALAR_NODE_CLASSES_BY_TAG:
         _SCALAR_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
+    elif tag["tag_uri"] in _UNIT_NODE_CLASSES_BY_TAG:
+        _UNIT_NODE_CLASSES_BY_TAG[tag["tag_uri"]].__doc__ = docstring
     else:
         _class_from_tag(tag, docstring)
 
@@ -575,5 +607,6 @@ for tag in _DATAMODELS_MANIFEST["tags"]:
 NODE_CLASSES = (
     list(_OBJECT_NODE_CLASSES_BY_TAG.values()) +
     list(_LIST_NODE_CLASSES_BY_TAG.values()) +
-    list(_SCALAR_NODE_CLASSES_BY_TAG.values())
+    list(_SCALAR_NODE_CLASSES_BY_TAG.values()) +
+    list(_UNIT_NODE_CLASSES_BY_TAG.values())
 )