+++ CHANGES.rst
@@ -1,7 +1,8 @@
 0.14.0 (unreleased)
 ===================
 
--
+- Explicitly add ``gwcs`` to the list of dependencies. [#108]
+- Add support for non-VOUnits to be used by Roman. [#109]
 
 0.13.0 (2022-08-23)
 ===================